Allow custom HID devices (not just mice)
This allows apps to specify custom USB HID devices, rather than just
mice.  This uses the functionality that was merged in
micropython/micropython#955.

I'm not super happy with the configuration here: there are 5 separate
new attributes that you can specify to control various aspects of the
HID behaviour: subclass (integer), protocol (integer), maximum packet
length (integer), polling interval (integer), and report descriptor (hex
string).  These correspond to the five values in the `hid` tuple that is
passed in pyb.usb_mode() call.

I didn't want to collapse the five values in to a single attribute
because I didn't want to get in to doing horrible parsing.  That said, I
think a case could be made for a USB-HID-tuple value of a JSON array
which is passed straight through to main.json.